Add object type metadata classes.
